Centralize more default test values
