+++ src/lib/filter/adjudicators/defaultTestObjects.ts
@@ -14,6 +14,7 @@ export const defaultFilters = {
 export const defaultBinding: Binding = {
   event: Event.ANY,
   filters: defaultFilters,
+  //   kind: { kind: "some-kind", group: "some-group" }, // Should it be this instead?? Used elsewhere.
   kind: modelToGroupVersionKind(kind.Pod.name),
   model: kind.Pod,
   isFinalize: false, //Lots of optionals that maybe don't belong here. Would be nice to choose to include
@@ -38,4 +39,25 @@ export const defaultKubernetesObject: KubernetesObject = {
   apiVersion: "some-version",
   kind: "some-kind",
   metadata: { name: "some-name" },
+};
+
+const callback = () => undefined;
+// const podKind = modelToGroupVersionKind(kind.Pod.name);
+const deploymentKind = modelToGroupVersionKind(kind.Deployment.name);
+const clusterRoleKind = modelToGroupVersionKind(kind.ClusterRole.name);
+
+export const groupBinding = {
+  callback,
+  event: Event.CREATE,
+  filters: defaultFilters,
+  kind: deploymentKind,
+  model: kind.Deployment,
+};
+
+export const clusterScopedBinding = {
+  callback,
+  event: Event.DELETE,
+  filters: defaultFilters,
+  kind: clusterRoleKind,
+  model: kind.ClusterRole,
 };
